+++ common/src/main/java/earth/terrarium/adastra/client/screens/PlanetsScreen.java
@@ -30,12 +30,14 @@ import net.minecraft.client.gui.components.Tooltip;
 import net.minecraft.client.gui.components.WidgetSprites;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.renderer.GameRenderer;
+import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.CommonComponents;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.Level;
@@ -43,6 +45,7 @@ import org.apache.commons.lang3.text.WordUtils;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
 
 public class PlanetsScreen extends AbstractContainerScreen<PlanetsMenu> {
@@ -95,7 +98,6 @@ public class PlanetsScreen extends AbstractContainerScreen<PlanetsMenu> {
 
     @Override
     protected void init() {
-
         super.init();
         buttons.clear();
         spaceStationButtons.clear();
@@ -112,6 +114,7 @@ public class PlanetsScreen extends AbstractContainerScreen<PlanetsMenu> {
         }
 
         backButton = addRenderableWidget(new LabeledImageButton(10, height / 2 - 85, 12, 12, BACK_BUTTON_SPRITES, b -> {
+            if (pageIndex != 2) this.scrollAmount = 0;
             pageIndex--;
             rebuildWidgets();
         }));
@@ -135,7 +138,10 @@ public class PlanetsScreen extends AbstractContainerScreen<PlanetsMenu> {
 
     private void createSolarSystemButtons() {
         selectedSolarSystem = null;
-        AdAstraData.solarSystems().forEach(solarSystem -> {
+
+        List<ResourceLocation> solarSystems = new ArrayList<>(AdAstraData.solarSystems());
+        solarSystems.sort(Comparator.comparing(ResourceLocation::getPath));
+        solarSystems.forEach(solarSystem -> {
             var button = addWidget(new LabeledImageButton(10, 0, 99, 20, BUTTON_SPRITES, b -> {
                 pageIndex = 1;
                 selectedSolarSystem = solarSystem;
@@ -163,7 +169,7 @@ public class PlanetsScreen extends AbstractContainerScreen<PlanetsMenu> {
 
     private void createSelectedPlanetButtons() {
         if (selectedPlanet == null) return;
-        var pos = menu.getLandingPos(selectedPlanet.dimension(), true);
+        BlockPos pos = menu.getLandingPos(selectedPlanet.dimension(), true);
         var button = addRenderableWidget(new LabeledImageButton(
             114, height / 2 - 77, 99, 20, BUTTON_SPRITES,
             b -> land(selectedPlanet.dimension()), ConstantComponents.LAND));
@@ -175,7 +181,7 @@ public class PlanetsScreen extends AbstractContainerScreen<PlanetsMenu> {
 
     private void addSpaceStationButtons(ResourceKey<Level> dimension) {
         menu.getOwnedAndTeamSpaceStations(dimension).forEach(station -> {
-            var pos = station.getSecond().position();
+            ChunkPos pos = station.getSecond().position();
             var button = addWidget(new LabeledImageButton(114, height / 2, 99, 20, BUTTON_SPRITES, b ->
                 landOnSpaceStation(dimension, pos), station.getSecond().name()));
             button.setTooltip(getSpaceStationLandTooltip(dimension, pos, station.getFirst()));
@@ -192,7 +198,7 @@ public class PlanetsScreen extends AbstractContainerScreen<PlanetsMenu> {
 
     public Tooltip getSpaceStationRecipeTooltip(ResourceKey<Level> planet) {
         List<Component> tooltip = new ArrayList<>();
-        var pos = menu.getLandingPos(planet, false);
+        BlockPos pos = menu.getLandingPos(planet, false);
         tooltip.add(Component.translatable("tooltip.ad_astra.construct_space_station_at", menu.getPlanetName(planet), pos.getX(), pos.getZ()).withStyle(ChatFormatting.AQUA));
 
         if (menu.isInSpaceStation(planet) || menu.isClaimed(planet)) {
@@ -223,6 +229,9 @@ public class PlanetsScreen extends AbstractContainerScreen<PlanetsMenu> {
         renderButtons(graphics, mouseX, mouseY, partialTick);
         backButton.visible = pageIndex > (hasMultipleSolarSystems ? 0 : 1);
         addSpaceStatonButton.visible = pageIndex == 2 && selectedPlanet != null;
+
+        // Prevent buttons from being pressed when outside view area.
+        buttons.forEach(button -> button.active = button.getY() > height / 2 - 63 && button.getY() < height / 2 + 88);
     }
 
     private void renderButtons(GuiGraphics graphics, int mouseX, int mouseY, float partialTick) {
@@ -382,11 +391,12 @@ public class PlanetsScreen extends AbstractContainerScreen<PlanetsMenu> {
     @Override
     public void onClose() {
         if (pageIndex > 0) {
+            if (pageIndex != 2) this.scrollAmount = 0;
             pageIndex--;
             rebuildWidgets();
             return;
         }
-        var player = menu.player();
+        Player player = menu.player();
         if (player.isCreative() || player.isSpectator()) super.onClose();
         else if (!(player.getVehicle() instanceof Rocket)) super.onClose();
     }